% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphRequest.R
\name{graphRequest}
\alias{graphRequest}
\title{Run a curl request on a neo4j graph}
\usage{
graphRequest(graph, endpoint, customrequest = c("POST", "GET"), postText)
}
\arguments{
\item{graph}{the neo4j connection}

\item{endpoint}{the endpoint for the request. To list all the available
endpoints:
\code{graphRequest(graph, endpoint="", customrequest="GET", postText="")$result}}

\item{customrequest}{the type of request: "POST" (default) or "GET"}

\item{postText}{the request body}
}
\value{
a list with the "header" and the "result" of the request (invisible)
}
\description{
Run a curl request on a neo4j graph
}
\seealso{
\code{\link[=startGraph]{startGraph()}} and \code{\link[=cypher]{cypher()}}
}
