% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_df.R
\name{import_from_df}
\alias{import_from_df}
\title{Imports a data.frame in the neo4j graph database}
\usage{
import_from_df(graph, cql, toImport, periodicCommit = 10000, ...)
}
\arguments{
\item{graph}{the neo4j connection}

\item{cql}{the CQL query to be applied on each row of toImport.
Use the 'row' prefix to refer to the data.frame column.}

\item{toImport}{the data.frame to be imported as "row".
Use "row.FIELD" in the cql query to refer to one FIELD of the toImport
data.frame}

\item{periodicCommit}{use periodic commit when loading the data
(default: 1000).}

\item{...}{further parameters for \code{\link[=cypher]{cypher()}}}
}
\description{
Imports a data.frame in the neo4j graph database
}
\seealso{
\code{\link[=cypher]{cypher()}}
}
