% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCql.R
\name{prepCql}
\alias{prepCql}
\title{Prepares a CQL query from a character vector}
\usage{
prepCql(...)
}
\arguments{
\item{...}{character vectors with cQL commands}
}
\value{
a well formated CQL query
}
\description{
Prepares a CQL query from a character vector
}
\examples{
prepCql(c(
 "MATCH (n)",
 "RETURN n"
))

}
\seealso{
\code{\link{readCql}}
}

