% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCql.R
\name{readCql}
\alias{readCql}
\title{Parse a CQL file and returned the prepared queries}
\usage{
readCql(file)
}
\arguments{
\item{file}{the name of the file to be parsed}
}
\value{
a character vector of well formated CQL queries
}
\description{
Parse a CQL file and returned the prepared queries
}
\seealso{
\code{\link[=cypher]{cypher()}} and \code{\link[=prepCql]{prepCql()}}
}
