% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startGraph.R
\name{startGraph}
\alias{startGraph}
\title{Prepare connection to neo4j database}
\usage{
startGraph(url, username = NA, password = NA, importPath = NA)
}
\arguments{
\item{url}{the DB url}

\item{username}{the neo4j user name
(default: NA; works only if authentication has been disabled in neo4j by
setting NEO4J.AUTH=none)}

\item{password}{the neo4j user password
(default: NA; works only if authentication has been disabled in neo4j by
setting NEO4J.AUTH=none)}

\item{importPath}{path to the import directory
(default: NA => no import directory). Import only works with local neo4j
instance.}
}
\value{
a connection to the graph DB:
a list with the url and necessary headers
}
\description{
Prepare connection to neo4j database
}
