% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startGraph.R
\name{startGraph}
\alias{startGraph}
\title{Prepare connection to neo4j database}
\usage{
startGraph(url, username, password)
}
\arguments{
\item{url}{the DB url}

\item{username}{the neo4j user name}

\item{password}{the neo4j user password}
}
\value{
a connection to the graph DB:
a list with the url and necessary headers
}
\description{
Prepare connection to neo4j database
}
